DEP-164 feat : 1년 유효기간 설정시 overflow로 인한 타입 변경
- Integer -> Long

+++ app/src/main/java/com/depromeet/threedays/front/support/TokenGenerator.java
@@ -16,7 +16,7 @@ public class TokenGenerator {
 	private String SECRET_KEY;
 
 	@Value("${security.jwt.token.validtime}")
-	private Integer TOKEN_VALID_TIME;
+	private Long TOKEN_VALID_TIME;
 
 	private static final String MEMBER_ID_CLAIM_KEY = "memberId";
